"Refactor motor control: move logic to device-specific struct
Removed motor control implementations from the `Motor` trait and migrated them to the `KDC101` device. Improved scaling functions with validation and range checks to ensure safety and handle non-finite numbers."
